chore: add src dir
